Added Amazon WorkMail Update Email Application
This application enables you to add a customized disclaimer and footer in the body of emails as they are being sent or received.

It is also a useful starting point for developers interested in making new Lambdas to update the content of in-transit email messages. See https://docs.aws.amazon.com/workmail/latest/adminguide/update-with-lambda.html and README.md for more details.